separate view and add folders
